git-svn-id: svn+ssh://scheduler.csc.calpoly.edu/repo@1007 13b268f7-8392-4c9d-bc7d-1e80a86bc2f9

+++ implementation/model/java/edu/calpoly/csc/scheduler/model/schedule/CourseDayAvail.java
@@ -12,7 +12,7 @@ import edu.calpoly.csc.scheduler.model.db.cdb.*;
 
 public class CourseDayAvail extends GenAvail<Course>
 {
-   private static final int serialVersionUID = 42;
+   private static final long serialVersionUID = 42;
 
    private BitSet ncoBitSet;
